requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==3.14.0
django-filter==23.3
Pillow==10.0.0



sqlparse==0.4.4


python-dotenv==1.0.0
